+++ zenoh-kotlin/src/commonMain/kotlin/io/zenoh/protocol/ZBytes.kt
@@ -249,7 +249,7 @@ import kotlin.reflect.typeOf
 class ZBytes internal constructor(internal val bytes: ByteArray) {
 
     companion object {
-        fun from(intoZBytes: IntoZBytes) = intoZBytes.into()
+        fun from(serializable: Serializable) = serializable.into()
         fun from(string: String) = ZBytes(string.toByteArray())
         fun from(byteArray: ByteArray) = ZBytes(byteArray)
         fun from(number: Number): ZBytes {
@@ -373,6 +373,71 @@ class ZBytes internal constructor(internal val bytes: ByteArray) {
         throw IllegalArgumentException("Unsupported type.")
     }
 
+}
+
+/**
+ * Serializable interface.
+ *
+ * Classes implementing this interface can be serialized into a ZBytes object.
+ *
+ * Example:
+ * ```kotlin
+ * class Foo(val content: String) : Serializable {
+ *
+ *   /*Inherits: Serializable*/
+ *   override fun into(): ZBytes = content.into()
+ * }
+ */
+interface Serializable {
+    fun into(): ZBytes
+}
+
+/**
+ * Deserializable interface.
+ *
+ * Classes implementing these two nested interfaces can be deserialized into a ZBytes object.
+ *
+ * The class must be declared as [Deserializable], but it's also necessary to make the companion
+ * object of the class implement the [Deserializable.From], as shown in the example below:
+ *
+ * ```kotlin
+ * class Foo(val content: String) : Deserializable {
+ *
+ *   companion object: Deserializable.From {
+ *      override fun from(zbytes: ZBytes): Foo {
+ *          return Foo(zbytes.toString())
+ *      }
+ *   }
+ * }
+ * ```
+ */
+interface Deserializable {
+    interface From {
+        fun from(zbytes: ZBytes): Serializable
+    }
+}
+
+@Throws
+fun Any?.into(): ZBytes {
+    return when (this) {
+        is String -> this.into()
+        is Number -> this.into()
+        is ByteArray -> this.into()
+        is Serializable -> this.into()
+        else -> throw IllegalArgumentException("Unsupported type")
+    }
+}
+
+fun Number.into(): ZBytes {
+    return ZBytes.from(this)
+}
+
+fun String.into(): ZBytes {
+    return ZBytes.from(this)
+}
+
+fun ByteArray.into(): ZBytes {
+    return ZBytes(this)
 }
 
 @PublishedApi
